Fixing Percentage Error And Output Training Bugs
Added a dense layer before the softmax which is necessary for stable output. Fixed a bug where I was printing the percentage of labels correct but it was titled as test error. Added a visual debug for predictions after the network training is complete. Added a time remaining estimate after each epoch when training.